getopt instead of manual parsing, style changes (#70)
* getopt instead of manual parsing, style changes

Something maybe unusual but I think it's better to use existing tools even here, instead of manually parsing the arguments, and I think it's more pleasing to the eye. Also, `getopt` checks for missing arguments for options which need one.

* Add stdio.h for fprintf

* Restore "vsync" option instead of "v"

That was a test, now it prints the correct (already) supported option.